requirements.txt
Django
django-bootstrap-v5
django-simple-captcha
django-ckeditor